refactor(teams): deprecate and remove team-based authentication endpoints
- Deprecated all team-based authentication endpoints in `api/src/api/http/teams.rs` by replacing implementations with 410 Gone responses.
- Moved the original team endpoints to `teams_old.rs` for reference during migration.
- Updated `core/src/types/team.rs` and `core/src/types/user.rs` to remove team user relations and mark team-related methods as deprecated.
- Updated user-related modules to use personal authentication instead of team-based.
- Changed application IDs from UUID to u32 in `api/src/api/http/applications.rs` for consistency.
- Fixed various user-related modules to improve error handling and data parsing.

This refactor removes legacy team-based authentication in favor of personal authentication, simplifying the codebase and improving maintainability.
